requirements.txt
google-generativeai>=0.8.0
mcp>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
python-dotenv>=1.0.0
gspread>=6.1.0
google-auth>=2.35.0
atlassian-python-api>=3.41.0
httpx>=0.27.0
aiofiles>=24.1.0
